Removed iframe logic. Caused too many warnings, and parsers essentially replace it.
Parsers are now being directly injected into the active tab.

Isolated the 'scraping' functionality into a single function that will be injected
into the active tab. It will either select the specialized parser if it exists, and
if necessary, fallback to a raw scraper.

Also added error-handling for unimplemented parsers, to gracefully select the raw
scraper instead.